+++ api_yamdb/api/serializers.py
@@ -2,6 +2,7 @@ import random
 import string
 
 from django.conf import settings
+from django.contrib.auth import get_user_model
 from django.core.mail import send_mail
 from django.db.models import Q
 from django.utils import timezone
@@ -14,13 +15,15 @@ from reviews.models import (
     ConfirmationCode,
     Genre,
     Title,
-    User,
-    UserNameValidator,
     Review,
     Comment
 )
 
+from users.models import UserNameValidator
+
 MIN_YEAR = settings.MIN_YEAR
+
+User = get_user_model()
 
 
 class BaseUserSerializer(serializers.ModelSerializer):
